% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorType.R
\name{Adduct.mz}
\alias{Adduct.mz}
\title{Calculate m/z}
\usage{
Adduct.mz(mass, adduct, charge)
}
\arguments{
\item{mass}{Molecule weight}

\item{adduct}{adduct mass}

\item{charge}{precursor charge value}
}
\value{
Returns the m/z value of the precursor ion
}
\description{
Calculate m/z
}

