% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{align_best.internal}
\alias{align_best.internal}
\title{Pick the best alignment result}
\usage{
align_best.internal(ref, peak, ms2.align, score.cutoff = 0.8, parent = NA)
}
\arguments{
\item{ref}{The identify metabolite ms2 matrix}

\item{peak}{The unknown metabolite ms2 matrix set}

\item{ms2.align}{Method for alignment of the ms2 matrix}
}
\value{
Returns the alignment result, which is a R list object with members:

    \code{list((ms2.matrix)ref, (ms2.matrix)candidate, score = [forward, reverse])}

    If the forward and reverse score cutoff less than score.cutoff, then this
    function will returns nothing.
}
\description{
Pick the best alignment result for unknown metabolite.
}

