% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/asciiView.R
\name{asciiView}
\alias{asciiView}
\title{ASCII View of Spectra Matrix}
\usage{
asciiView(spectra, into.cutoff = 0.05, maxWidth = 80,
  show.mz.labels = TRUE)
}
\arguments{
\item{spectra}{The spectra matrix should contains at least two column:
first column for \code{m/z} and second column for \code{intensity}.}

\item{into.cutoff}{All of the \code{m/z} fragment which its intensity
value less than this percetange cutoff, will be display on the output}

\item{maxWidth}{The max width of the text viewer in ncharacters.}
}
\value{
The text of the spectra viewer result, you should use \code{cat}
  function for display this \code{spectra} on the console
}
\description{
ASCII View of Spectra Matrix
}

