% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{create_filter.skips}
\alias{create_filter.skips}
\title{Create skips handler for KEGG id}
\usage{
create_filter.skips(kegg_id.skips, debug.echo = TRUE)
}
\arguments{
\item{kegg_id.skips}{A character vector that contains KEGG id
want to ignores in the metaDNA prediction process.}
}
\value{
This function returns a lambda function that can determine the
  given kegg id vector which is not in the input \code{kegg_id.skips}.
}
\description{
Create skips handler for KEGG id
}
