% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_handler.R
\name{Delete.EmptyKEGG}
\alias{Delete.EmptyKEGG}
\title{Removes empty \code{KEGG} row}
\usage{
Delete.EmptyKEGG(dataframe, col.name = "KEGG")
}
\arguments{
\item{d}{A \code{dataframe} object which contains the metabolite
annotation result data. This dataframe must contains a column
which is named \code{KEGG}.}
}
\value{
A subset of the input \code{dataframe} with all KEGG column
    value non-empty.
}
\description{
Removes all of the rows in a given \code{dataframe}
     which its \code{KEGG} column value is string empty.
}
