% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA_iteration.R
\name{do.Predicts}
\alias{do.Predicts}
\title{do metaDNA prediction for given KEGG compounds}
\usage{
do.Predicts(
  KEGG_cpd,
  identified,
  KEGG.partners,
  unknown.query,
  unknown,
  do.align,
  score.cutoff
)
}
\arguments{
\item{KEGG_cpd}{the kegg compound id}

\item{do.align}{a lambda function for evaluate align score of two spectra matrix.}
}
\description{
do metaDNA prediction for given KEGG compounds
}
