% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{do.write.network}
\alias{do.write.network}
\title{Write metaDNA infer network}
\usage{
do.write.network(write, infer)
}
\arguments{
\item{write}{The text file stream handler}

\item{infer}{The metaDNA infer network for the resulted unknown metabolite feature identification.}
}
\description{
Write metaDNA infer network
}
