% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match.R
\name{.PrecursorType.MatchImpl}
\alias{.PrecursorType.MatchImpl}
\title{Calculate for each precursor type, and then returns the
min tolerance type as the match result}
\usage{
.PrecursorType.MatchImpl(mass, precursorMZ, charge, chargeMode, tolerance,
  debug.echo)
}
\description{
Calculate for each precursor type, and then returns the
min tolerance type as the match result
}
