% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA_seeds.R
\name{extends.seeds}
\alias{extends.seeds}
\title{Convert output as metaDNA seeds}
\usage{
extends.seeds(output)
}
\description{
The identify list only provides ms2 spectra matrix and KEGG id.
  KEGG id is the names of the identify list.
  So, the identify list object its structure looks like: 

  \code{
     identify \{
        KEGG_id1 => list(spectra),
        KEGG_id2 => list(spectra),
        ...
     \}
  }
}

