% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA_seeds.R
\name{extends.seeds.top}
\alias{extends.seeds.top}
\title{Subset of the seeds by top scores}
\usage{
extends.seeds.top(seeds, n = 5)
}
\arguments{
\item{seeds}{The seeds data collection}

\item{n}{top n, by default is top 5}
}
\value{
The subset of the input seeds collection.
}
\description{
Subset of the seeds by top scores
}

