% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelKEGG.R
\name{find.KEGG}
\alias{find.KEGG}
\title{Search KEGG meta database by MS1}
\usage{
find.KEGG(precursor, KEGG, tolerance = tolerance.ppm(),
  precursor_type = "[M+H]+")
}
\arguments{
\item{precursor}{The MS1 information, \code{m/z}}

\item{KEGG}{A list object with data structure like:
\code{DB[KEGGID] = [KEGGID, mass, formula, name]}}

\item{tolerance}{The tolerance predicate method, by default is a method with ppm 20 precision.}

\item{precursor_type}{The \code{precursor_type} of this \code{MS1} value.}
}
\value{
Returns one of the element in list \code{KEGG}
}
\description{
Search KEGG meta database by MS1
}

