% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_handler.R
\name{get.kegg_precursorMZ}
\alias{get.kegg_precursorMZ}
\title{Get precursor m/z data list}
\usage{
get.kegg_precursorMZ(KEGG_meta, precursor_type, mode)
}
\arguments{
\item{KEGG_meta}{The kegg meta dataset that loaded from the metaDNA package}

\item{precursor_type}{A character vector data for get m/z from the precursor data matrix}

\item{mode}{libtype value, value could be 1 or -1}
}
\description{
Get precursor m/z data list
}
