% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorType.R
\name{get.mass}
\alias{get.mass}
\title{Get mass calculator}
\usage{
get.mass(chargeMode, charge, PrecursorType)
}
\arguments{
\item{chargeMode}{Character value of \code{+/-}.}
}
\description{
Get mass calculator
}

