% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get.PrecursorMZ.Auto}
\alias{get.PrecursorMZ.Auto}
\title{Calculate m/z}
\usage{
get.PrecursorMZ.Auto(M, precursorType)
}
\arguments{
\item{M}{Molecule mass}

\item{precursorType}{Charge mode was parsed from this parameter string,
it is convenient, but inefficient when in batch mode.}
}
\value{
-1 means target precursor type is not found.
}
\description{
Calculate \code{m/z} for mass by given precursor type
}

