% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorType.R
\name{get.PrecursorMZ}
\alias{get.PrecursorMZ}
\title{Calculate m/z for M by given precursor type}
\usage{
get.PrecursorMZ(M, precursorType)
}
\value{
-1 means target precursor type is not found.
}
\description{
Calculate m/z for M by given precursor type
}

