% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{getPolarity}
\alias{getPolarity}
\title{Get ionlization mode}
\usage{
getPolarity(type)
}
\arguments{
\item{type}{The precursor type name, it should be in format like: \code{[M+H]+}.}
}
\value{
Function returns character \code{+} or \code{-}.
}
\description{
Get ionlization mode from a given precursor type name
}

