% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{kegg.match.handler}
\alias{kegg.match.handler}
\title{Match unknown by mass}
\usage{
kegg.match.handler(meta.KEGG, precursor_type = "[M+H]+",
  tolerance = tolerance.ppm(20))
}
\description{
Match unknown by mass
}

