% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGG_handler.R
\name{kegg.match}
\alias{kegg.match}
\title{Match unknown feature from given KEGG partners}
\usage{
kegg.match(partners_id, kegg.mass, kegg.ids, kegg.mz, kegg.list, precursor_type,
  unknown.mz, tolerance)
}
\arguments{
\item{partners_id}{The kegg partner compound id list}
}
\description{
Match unknown feature from given KEGG partners
}

