% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGHandler.R
\name{kegg.partners}
\alias{kegg.partners}
\title{Find kegg reaction partner}
\usage{
kegg.partners(kegg_id)
}
\arguments{
\item{kegg_id}{The kegg compound id of the identified compound.}
}
\value{
A kegg id vector which is related to the given \code{kegg_id}.
}
\description{
Find KEGG reaction partner compound based on the kegg
    reaction definition.
}

