% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{metaDNA.impl}
\alias{metaDNA.impl}
\title{Try to annotate unknown metabolite}
\usage{
metaDNA.impl(
  unknown.query,
  identify.ms2,
  unknown,
  ms2.align,
  trace,
  score.cutoff = 0.8
)
}
\arguments{
\item{identify.ms2}{The identify metabolite's MS/MS matrix data.}

\item{unknown}{Unknown metabolite's peaktable and peak_ms2 data.}

\item{ms2.align}{The ms2 alignment method, this function method should returns \code{forward}
     and \code{reverse} alignment score result list which its data structure in format like:

     \code{list(forward = score1, reverse = score2)}.}

\item{score.cutoff}{SSM algorithm alignment score cutoff, by default is \code{0.8}. The unknown
metabolite which its forward and reverse alignment score greater than this cutoff value both,
will be picked as the candidate result.}

\item{KEGG.partners}{Related to the identified KEGG id based on the kegg reaction definitions.
Using for find unknown metabolite ms2 data.}
}
\value{
A set of unknown identify result based on the given related kegg partners id set.
}
\description{
Try to annotate the unknown metabolite as a given set of
    kegg metabolite candidates. The ms2 alignment is based on the identified
    metabolite ms2 data.
}
\details{
Algorithm routine:

   \code{\cr
         KEGG.partners -> kegg.match.handler\cr
                       -> unknown index\cr
                       -> unknown ms2\cr
                       -> identify.ms2 alignment\cr
                       -> is similar?\cr
   }

   \enumerate{
       \item yes, identify the unknown as \code{\link{kegg.partners}}
       \item no, returns \code{NULL}
   }

   One identify metabolite have sevral kegg partners based on the metabolism network definition
   So, this function find every partner in unknown, and returns a set of unknown identify result
   But each unknown identify only have one best identify ms2 alignment result.
}
