% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{metaDNA.iteration}
\alias{metaDNA.iteration}
\title{Run a metaDNA prediction iteration}
\usage{
metaDNA.iteration(identify, kegg.partners, filter.skips, unknown, do.align,
  match.kegg, score.cutoff)
}
\arguments{
\item{identify}{The seeds data for the metaDNA algorithm.}

\item{kegg.partners}{A lambda function for find reaction partners for a given list of KEGG compound id.}

\item{unknown}{The user sample data}

\item{do.align}{A lambda function that provides spectra alignment}
}
\description{
Run a metaDNA prediction iteration
}
\details{
The \code{do.align} function should take two parameter:
    The spectra matrix of query and reference and retuns a score vector
    which produced by forward and reverse spectra alignment.
}

