% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{metaDNA}
\alias{metaDNA}
\title{Identify unknown metabolite by metaDNA algorithm}
\usage{
metaDNA(identify, unknown, do.align, precursor_type = c("[M+H]+",
  "[M]+"), tolerance = assert.deltaMass(0.3), rt.adjust = function(rt,
  KEGG_id) 1, score.cutoff = 0.8, kegg_id.skips = NULL,
  seeds.all = TRUE, seeds.topn = 5, iterations = 20,
  network.class_links = NULL)
}
\arguments{
\item{identify}{The metaDNA seeds data input, A \code{list} object with two members

     \code{identify = list((data.frame) meta.KEGG, (list) peak_ms2)}

     where:

        \code{meta.KEGG} is a data.frame object that should contains
              a KEGG id column, and a index column to read \code{peak_ms2} list.

        \code{peak_ms2} is a MS/MS list, MS/MS matrix should contains at
              least two column: \code{mz} and \code{into}}

\item{unknown}{A \code{list} object with two members:

     \code{unknown = list((data.frame) peaktable, (list) peak_ms2)}

     where:

     \code{peaktable} is a data.frame object that contains the MS1
     information: \code{mz} and \code{rt}, and a additional column index
     to read \code{peak_ms2} list.

     \code{peak_ms2} is a MS/MS list, MS/MS matrix should contains at
     least two column: \code{mz} and \code{into}}

\item{do.align}{The MS/MS alignment method, which is in format like: \code{function(q, s)}
Where \code{q} and \code{s} is a ms2 spectra matrix. Due to the reason of the \code{q}
and \code{s} spectra data is comes from the same sample and same instrument, so that
using the \code{SSM} spectra alignment algorithm for this parameter is recommended.}

\item{precursor_type}{By default is positive mode with the \code{H+} adduct for
search unknown metabolites.}

\item{tolerance}{KEGG compound match with this tolerance, m/z equalient compares tolerance,
by default is less than ppm 20.}

\item{rt.adjust}{The rt adjust lambda function, it takes two parameter:
\enumerate{
    \item \code{rt}, the rt of unidentified sample ms1 feature,
    \item \code{KEGG_id} the KEGG_cpd that assigned to current unidentified sample ms1 feature.
}
and then returns a rt alignment score. If this parameter is ignored,
then will use score 1 as default means no rt adjustment score.}

\item{score.cutoff}{MS/MS similarity cutoff for identify ms2 alignment with unknown ms2}

\item{kegg_id.skips}{You can put the kegg compound ids in this character vector
If you don't want some specific metabolite was indeified from this
metaDNA algorithm function.}
}
\value{
A \code{identify} parameter data structure like metabolite identify
     result for \code{unknown} parameter input data
}
\description{
How to: The basic idea of the \code{MetaDNA} algorightm is
     using the identified ms2 data to align the unknown ms2 data.
}
\details{
Algorithm implementation and details see: \code{\link{metaDNA.impl}}.
   The ms2 matrix should be in format like:

    \code{
        mz into\cr
        xxx xxx\cr
        xxx xxx\cr
        xxx xxx\cr
    }
}
