% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MolWeight.R
\name{MolWeight}
\alias{MolWeight}
\title{Molecule weight calculate helper}
\usage{
MolWeight()
}
\value{
A list module that contains member function:
   \enumerate{
   \item \code{Eval}
   \item \code{Weight}
}
}
\description{
Molecule weight calculate helper
}

