% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelKEGG.R
\name{ms2.similar.fn}
\alias{ms2.similar.fn}
\title{The default SSM \code{ms/ms} similarity method}
\usage{
ms2.similar.fn(threshold = 0.8)
}
\arguments{
\item{threshold}{The SSM score threshold, by default is greater than 0.8}
}
\value{
returns a lambda function that using for predicate is query ms2 matrix
        similar to the subject ms2 matrix.
}
\description{
The default SSM \code{ms/ms} similarity method
}

