% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/MzCalculator.R
\name{mz.calculator}
\alias{mz.calculator}
\title{Calculate m/z matrix}
\usage{
mz.calculator(mass, mode = c(1, -1), debug = FALSE)
}
\arguments{
\item{mass}{The molecule weight.}

\item{mode}{The ionization mode, by default is positive mode.
\enumerate{
   \item  1: \code{positive}
   \item -1: \code{negative}
}.}

\item{debug}{A logical flag to indicate that this function should
output debug info onto the console screen?}
}
\value{
\code{m/z} dataframe for all precursor types.
}
\description{
Calculate all of the m/z value by enumerate all of the precursor types.
}
