% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{network.trace}
\alias{network.trace}
\title{Network Visualize of the MetaDNA result}
\usage{
network.trace(result)
}
\arguments{
\item{result}{The output result data set which is generated by \code{metaDNA} function.}
}
\description{
Network Visualize of the MetaDNA result
}
