% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/algorithm.R
\name{pickbest.internal}
\alias{pickbest.internal}
\title{Pick best alignment result for single ms1 feature}
\usage{
pickbest.internal(align)
}
\arguments{
\item{align}{The alignment result of the ms2 spectra that align with reference spectra}
}
\description{
due to the reason of all of the ms2 spectra in current one peak pack
is corresponding to one ms1 feature, so that the precursor m/z and rt 
all the same. so no needs for rt.adjust score at here

we just required compare the best dot product score at here
}
