% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mgf.R
\name{read.mgf}
\alias{read.mgf}
\title{Read a given mgf file}
\usage{
read.mgf(fileName)
}
\value{
Returns a list of mgf ions that parsed from the given mgf spectrum data file.
}
\description{
Read a given mgf file
}
