% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{removes.empty_KEGG}
\alias{removes.empty_KEGG}
\title{Removes empty \code{KEGG} row}
\usage{
removes.empty_KEGG(d)
}
\arguments{
\item{d}{A \code{dataframe} object which contains the metabolite
annotation result data. This dataframe must contains a column
which is named \code{KEGG}.}
}
\value{
A subset of the input \code{dataframe} with all KEGG column
    value non-empty.
}
\description{
Removes all of the rows in a given \code{dataframe}
     which its \code{KEGG} column value is string empty.
}

