% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorType.R
\name{Reverse.mass}
\alias{Reverse.mass}
\title{Calculate mass from m/z}
\usage{
Reverse.mass(precursorMZ, M, charge, adduct)
}
\arguments{
\item{precursorMZ}{MS/MS precursor adduct ion m/z}

\item{M}{The number of the molecule for formula a precursor adduct ion.}

\item{charge}{Net charge of the ion}

\item{adduct}{Adduct mass}
}
\value{
The molecule mass.
}
\description{
Calculate the molecule mass from precursor adduct ion m/z
}

