% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{save.network}
\alias{save.network}
\title{Save metaDNA network}
\usage{
save.network(infer, outputdir, fileName = "MetaDNA.Xml")
}
\description{
Save metaDNA network
}
\details{
save network data into Xml file for export data for network visualization.
}
