% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance.deltaMass}
\alias{tolerance.deltaMass}
\title{Tolerance in Mass delta mode}
\usage{
tolerance.deltaMass(da = 0.3)
}
\arguments{
\item{da}{The mass delta value. By default if two mass value which
their delta value is less than \code{0.3da}, then
the predicate will be true.}
}
\value{
Function returns a lambda function that can be using for
        tolerance predication.
}
\description{
Tolerance in Mass delta mode
}
