% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tolerance.R
\name{tolerance}
\alias{tolerance}
\title{\code{m/z} tolerance helper}
\usage{
tolerance(threshold = 0.3, method = c("da", "ppm"))
}
\value{
This function returns a list object with members:
   \enumerate{
       \item \code{threshold} The tolerance threshold.
       \item \code{method} The tolerance method name, which could be \code{da} or \code{ppm}.
       \item \code{assert} The tolerance assert method function, which could be
             \code{\link{assert.deltaMass}} or \code{\link{assert.ppm}} based on the \code{method}
             parameter value.
       \item \code{is.low.resolution} The tolerance \code{assert} method is a low resolution
             method? \code{da} method is \code{TRUE}, and \code{ppm} method is \code{FALSE}.
   }
}
\description{
\code{m/z} tolerance helper
}
