% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/debugger.R
\name{trace.node}
\alias{trace.node}
\title{A unified method for create node ID}
\usage{
trace.node(seed)
}
\arguments{
\item{seed}{The identified metaDNA seed data.}
}
\description{
A unified method for create node ID
}
\details{
Create debug data of metaDNA network elongation.
   The input seed parameter should have such data structure:
   \enumerate{
      \item \code{feature} is the ms1 feature ID
      \item \code{KEGG} is the kegg compound id that current seed object identified as
      \item \code{ref} is the ms2 spectrum matrix reference index value, which should contains two member: \code{file} and \code{scan}
   }
}
