% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA_seeds.R
\name{extends.seeds}
\alias{extends.seeds}
\title{Convert output as metaDNA seeds}
\usage{
extends.seeds(
  output,
  rt.adjust = function(rt, KEGG_id) 1,
  seeds.all = TRUE,
  seeds.topn = 5
)
}
\arguments{
\item{seeds.all}{If this parameter is enable, then all of the hit from the 
alignment result will be used as metaDNA seeds in the next iteration, 
otherwise the best hit will be picked from the alignment result as 
the seeds if this parameter is set to \code{FALSE}
But if too many seeds in a cluster, then will caused the dataset is too large,
And this will makes the alignment iteration took very long long time for run. 
So just pick the top 5 result when requires all alignment hit as seeds.}
}
\description{
The identify list only provides ms2 spectra matrix and KEGG id.
  KEGG id is the names of the identify list.
  So, the identify list object its structure looks like: 

  \code{
     identify \{
        KEGG_id1 => list(spectra),
        KEGG_id2 => list(spectra),
        ...
     \}
  }
}
