% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get.mass}
\alias{get.mass}
\title{Get mass calculator}
\usage{
get.mass(chargeMode, PrecursorType)
}
\arguments{
\item{chargeMode}{Character value of \code{+/-}.}

\item{PrecursorType}{The precursor type full name or brief name.}
}
\value{
Returns a function for calculate mass from \code{m/z} value.
}
\description{
Get mass calculator
}
