% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Utils.R
\name{get.PrecursorMZ}
\alias{get.PrecursorMZ}
\title{Calculate m/z}
\usage{
get.PrecursorMZ(M, precursorType)
}
\arguments{
\item{M}{Molecule mass}
}
\value{
-1 means target precursor type is not found.
}
\description{
Calculate \code{m/z} for mass by given precursor type
}
