% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{kegg.match.handler}
\alias{kegg.match.handler}
\title{Match unknown by mass}
\usage{
kegg.match.handler(meta.KEGG, unknown.mz, precursor_type = "[M+H]+",
  tolerance = tolerance.ppm(20))
}
\arguments{
\item{unknown.mz}{This mz parameter is a \code{m/z} vector from the \code{unknown$peaktable}}
}
\value{
Returns the index vector in \code{unknown.mz} vector.
}
\description{
Match unknown by mass
}

