% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/KEGGHandler.R
\name{kegg.match.handler}
\alias{kegg.match.handler}
\title{Match unknown by mass}
\usage{
kegg.match.handler(unknown.mz, precursor_type = c("[M+H]+", "[M]+"),
  tolerance = assert.deltaMass(0.3))
}
\arguments{
\item{unknown.mz}{This mz parameter is a \code{m/z} vector from the
\code{unknown$peaktable}}

\item{precursor_type}{precursor type that using for calculate
\code{m/z} from KEGG metabolite mass value.}
}
\value{
Returns the index vector in \code{unknown.mz} vector.
}
\description{
Match unknown by mass
}
\details{
\code{mass - mz - kegg_id}
}

