% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ModelKEGG.R
\name{KEGG.rxnNetwork}
\alias{KEGG.rxnNetwork}
\title{ä½¿ç”¨KEGGçš„ä»£è°¢ç‰©æ•°æ®åº“å’Œä»£è°¢ååº”è¿‡ç¨‹æ•°æ®åº“æ‰¾å‡ºå’Œç›®æ ‡å·²ç»é‰´å®šå‡ºçš„ä»£è°¢ç‰©çš„æ‰€æœ‰ä»£è°¢è¿‡ç¨‹ç›¸å…³çš„æœªé‰´å®šä»£è°¢ç‰©çš„KEGGæ³¨é‡Šä¿¡æ¯}
\usage{
KEGG.rxnNetwork(identified, sample, KEGG, KEGG.rxn,
  ms2.similar = ms2.similar.fn(0.65), tolerance = tolerance.ppm(),
  precursor_type = "[M+H]+")
}
\arguments{
\item{identified}{The identified metabolite annotation \code{data.frame} 
result set. This result set should contains a column data which is 
named \code{KEGG}.}

\item{sample}{ä¸€èˆ¬æ˜¯æœªé‰´å®šçš„ä»£è°¢ç‰©ç»“æœï¼Œpeak_ms2ï¼<U+FFFD>?}

\item{KEGG}{KEGGçš„ä»£è°¢ç‰©æ³¨é‡Šä¿¡æ¯åº“ï¼Œå¯ä»¥å°†metadbä¹‹ä¸­çš„å«æœ‰KEGGç¼–å·çš„ç‰©è´¨æ³¨é‡Šä¿¡æ¯å–å­é›†}

\item{ms2.similar}{æ¯”è¾ƒä¸¤ä¸ªäºŒçº§è´¨è°±çŸ©é˜µæ˜¯å¦ç›¸ä¼¼ï¼Œå‡½æ•°è¿”å›é€»è¾‘å€<U+FFFD>

  å‡è®¾äºŒçº§è´¨è°±çš„ç›¸ä¼¼åº¦çš„æ¯”å¯¹å‡½æ•°çš„æ¥å£å½¢å¼ä¸ºï¼š
  ms2.alignment <- function(q, s) { score = SSM(q, s)ï¼<U+FFFD> score >= threshold; }

q,sä¸ºäºŒçº§çŸ©é˜µï¼š

mz, into
mz, into
mz, into
mz, into}

\item{RXN}{KEGGæ•°æ®åº“ä¹‹ä¸­çš„ä»£è°¢è¿‡ç¨‹çš„é›†å<U+FFFD>}
}
\description{
ä½¿ç”¨KEGGçš„ä»£è°¢ç‰©æ•°æ®åº“å’Œä»£è°¢ååº”è¿‡ç¨‹æ•°æ®åº“æ‰¾å‡ºå’Œç›®æ ‡å·²ç»é‰´å®šå‡ºçš„ä»£è°¢ç‰©çš„æ‰€æœ‰ä»£è°¢è¿‡ç¨‹ç›¸å…³çš„æœªé‰´å®šä»£è°¢ç‰©çš„KEGGæ³¨é‡Šä¿¡æ¯
}

