% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{metaDNA.iteration}
\alias{metaDNA.iteration}
\title{Run a metaDNA prediction iteration}
\usage{
metaDNA.iteration(
  identify,
  filter.skips,
  unknown,
  do.align,
  match.kegg,
  score.cutoff,
  network
)
}
\arguments{
\item{identify}{The seeds data for the metaDNA algorithm.}

\item{unknown}{The user sample data}

\item{do.align}{A lambda function that provides spectra alignment}

\item{kegg.partners}{A lambda function for find reaction partners for a given list of
KEGG compound id.}

\item{unknow.matches}{function evaluate result of \code{\link{kegg.match.handler}},
this function descript that how to find out the unknown metabolite from a given
set of identify related kegg partners compound id set.}
}
\description{
Run a metaDNA prediction iteration
}
\details{
The \code{do.align} function should take two parameter:
    The spectra matrix of query and reference and retuns a score vector
    which produced by forward and reverse spectra alignment.
}
