% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metaDNA.R
\name{metaDNA}
\alias{metaDNA}
\title{Identify unknown metabolite by metaDNA algorithm}
\usage{
metaDNA(identify, unknown, meta.KEGG, ms2.align, precursor_type = "[M+H]+",
  tolerance = tolerance.ppm(20))
}
\arguments{
\item{unknown}{A \code{list} object with two members:

     \code{unknown = list((data.frame) peaktable, (list) peak_ms2)}

     where:

     \code{peaktable} is a data.frame object that contains the MS1
     information: \code{mz} and \code{rt}, and a additional column index
     to read \code{peak_ms2} list.

     \code{peak_ms2} is a MS/MS list, MS/MS matrix should contains at
     least two column: \code{mz} and \code{into}}

\item{precursor_type}{By default is positive mode with the \code{H+} adduct for
search unknown metabolites.}
}
\value{
A \code{identify} parameter data structure like metabolite identify
     result for \code{unknown} parameter input data
}
\description{
How to: The basic idea of the \code{MetaDNA} algorightm is
     using the identified ms2 data to align the unknown ms2 data.
}
\details{
Algorithm implementation and details see: \code{\link{metaDNA.impl}}.
}

