% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/PrecursorType.R
\name{PrecursorType.Match}
\alias{PrecursorType.Match}
\title{Match the precursor type}
\usage{
PrecursorType.Match(mass, precursorMZ, charge, chargeMode = "+",
  tolerance = tolerance.deltaMass(0.3), debug.echo = TRUE)
}
\arguments{
\item{mass}{Molecular mass}

\item{precursorMZ}{Precursor m/z value of the ion.}

\item{charge}{The charge value of the ion}

\item{tolerance}{Tolerance between two mass value, by default is 0.3 da,
   if this parameter is a numeric value, then means tolerance by ppm value.
   There are two pre-defined tolerance function:

   \enumerate{
       \item \link{\code{tolerance.deltaMass}}
       \item \link{\code{tolerance.ppm}}
   }}
}
\description{
Match the precursor type through min ppm value match.
}
\examples{
mass = 853.33089

find.PrecursorType(853.33089, 307.432848,  charge = 3) # pos "[M+3Na]3+" charge = 3,  307.432848
find.PrecursorType(853.33089, 1745.624938, charge = 1) # pos "[2M+K]+"   charge = 1,  1745.624938
find.PrecursorType(853.33089, 854.338166,  charge = 1) # pos "[M+H]+"    charge = 1,  854.338166

find.PrecursorType(853.33089, 283.436354,  charge = -3, chargeMode = "-") # neg "[M-3H]3-"  charge = -3, 283.436354
find.PrecursorType(853.33089, 2560.999946, charge = -1, chargeMode = "-") # neg "[3M-H]-"   charge = -1, 2560.999946
find.PrecursorType(853.33089, 852.323614,  charge = -1, chargeMode = "-") # neg "[M-H]-"    charge = -1, 852.323614

}

